feat: Add total_upvotes and total_comments columns to post and comment tables
This commit adds the total_upvotes column to the post table and the total_comments column to the comment table. The columns are initialized with a default value of 0 and are set to not null. This change allows for tracking the total number of upvotes and comments associated with a post or comment.