refactor: Update application.properties for staging environment
Update the spring.config.import property in application.properties to use the .env.staging file instead of .env.development. This change ensures that the staging environment is properly configured. Additionally, remove the static.resource.url property and replace it with storage.bucket to align with the updated configuration metadata. Finally, disable livereload and restart in the devtools section to prevent unnecessary reloading and restarting in the staging environment.